fix: only apply necessary build args to extending services
